Revert "refactor: resolve remaining codescene issues"
This reverts commit cb99347fd5cf7256b54bfeb59edcf28ff6ade9f7.
