+++ src/components/common/RecentlyViewedEvents.jsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useCallback } from 'react';
+import { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import useRecentlyViewed from '../../hooks/useRecentlyViewed';
 import LazyImage from './LazyImage';
@@ -155,9 +155,11 @@ export const RecentlyViewedTracker = ({ event }) => {
 };
 
 /**
- * Custom hook encapsulating state and interaction logic for the RecentlyViewedEvents component.
+ * RecentlyViewedEvents Component
+ *
+ * Displays a horizontal scrollable strip of recently viewed events.
  */
-const useRecentlyViewedEventsController = ({ maxVisible, onEventClick }) => {
+const RecentlyViewedEvents = ({ maxVisible = 6, onEventClick }) => {
   const { recentlyViewed, removeRecentlyViewed, clearHistory } = useRecentlyViewed();
   const navigate = useNavigate();
   const [showAll, setShowAll] = useState(false);
@@ -170,59 +172,26 @@ const useRecentlyViewedEventsController = ({ maxVisible, onEventClick }) => {
     return () => clearTimeout(timer);
   }, [confirmClear]);
 
+  if (recentlyViewed.length === 0) return null;
+
   const visibleEvents = showAll ? recentlyViewed : recentlyViewed.slice(0, maxVisible);
 
-  const handleCardClick = useCallback((event) => {
+  const handleCardClick = (event) => {
     if (onEventClick) {
       onEventClick(event);
     } else {
       navigate(`/events/${event.id}`);
     }
-  }, [onEventClick, navigate]);
+  };
 
-  const handleClear = useCallback(() => {
+  const handleClear = () => {
     if (confirmClear) {
       clearHistory();
       setConfirmClear(false);
     } else {
       setConfirmClear(true);
     }
-  }, [confirmClear, clearHistory]);
-
-  const handleToggleShowAll = useCallback(() => {
-    setShowAll((v) => !v);
-  }, []);
-
-  return {
-    recentlyViewed,
-    visibleEvents,
-    showAll,
-    confirmClear,
-    handleCardClick,
-    handleClear,
-    handleToggleShowAll,
-    removeRecentlyViewed,
   };
-};
-
-/**
- * RecentlyViewedEvents Component
- *
- * Displays a horizontal scrollable strip of recently viewed events.
- */
-const RecentlyViewedEvents = ({ maxVisible = 6, onEventClick }) => {
-  const {
-    recentlyViewed,
-    visibleEvents,
-    showAll,
-    confirmClear,
-    handleCardClick,
-    handleClear,
-    handleToggleShowAll,
-    removeRecentlyViewed,
-  } = useRecentlyViewedEventsController({ maxVisible, onEventClick });
-
-  if (recentlyViewed.length === 0) return null;
 
   return (
     <section className="recently-viewed-section" aria-label="Recently Viewed Events">
@@ -230,7 +199,7 @@ const RecentlyViewedEvents = ({ maxVisible = 6, onEventClick }) => {
         count={recentlyViewed.length}
         maxVisible={maxVisible}
         showAll={showAll}
-        onToggleShowAll={handleToggleShowAll}
+        onToggleShowAll={() => setShowAll((v) => !v)}
         confirmClear={confirmClear}
         onClear={handleClear}
       />
